asdf

+++ src/main/java/com/example/noticeboardprj/repository/UserRepository.java
@@ -3,7 +3,7 @@ package com.example.noticeboardprj.repository;
 import com.example.noticeboardprj.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<UserEntity, Integer> {
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
 
     Boolean existsByUsername(String username);
